레베2 과제 요구사항 8완료-3

+++ MultiplyOperator.java
@@ -1,7 +1,9 @@
 package Work;
 
 public class MultiplyOperator {
-    public double operate(double a, double b) {
+
+
+    public double operate(int a, int b) {
         return a*b;
     }
